algorithm_2_11_2022/requirements.txt
autopep8==2.0.0
pycodestyle==2.9.1
pyserial==3.5
tomli==2.0.1
